feat!: add streaming MP4 support
Add support for streaming MP4 (H.264) over HTTP.

The distinction between API and format is removed, and instead only a
format is used (which is then restricting API, videocodec, and
container).  So, you no longer should add either `videocodec` or
`container`, as these will be overwritten by the format choice.

Explicit refresh when changing format has been removed as well, since
this causes an unnecessary extra fetch.  Instead, an explicit cache bust
is added to the still image when mounting the component.

Because there might be CORS issues when fetching over HTTP directly from
a camera via the development server, the `dev` command was modified so
it takes a camera IP via the `MSP_CAMERA` environment variable. The
webpack dev server then proxies video requests to that IP. This means we
no longer need the hostname input, as this doesn't work with all formats
(there is a short section now in the README.md about CORS).
Note that this does not apply to the multistream example, where you will
still need to proxy all cameras yourself if using the streaming MP4
format.

BREAKING CHANGES:
 - `format` is now one of 'RTP_H264', 'RTP_JPEG', 'JPEG', or 'MP4_H264'
 - `api` as a property has been removed from `PlaybackArea` and
   `Settings` and is replaced by the `format` property
 - the refresh key is no longer increased on a format change, and
   instead the cache is busted only for still images